# hotstock-be-v1/src/modules/articles/articles.controller.ts
import {
  Controller,
  Get,
  Post,
  Patch,
  Delete,
  Param,
  Body,
  Query,
  Req,
  UseGuards,
} from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
  ApiQuery,
} from '@nestjs/swagger';
import { Role } from '@prisma/client';
import { FastifyRequest } from 'fastify';
import { ArticlesService } from './articles.service';
import { CreateArticleDto } from './dto/create-article.dto';
import { UpdateArticleDto } from './dto/update-article.dto';
import { ArticleListQueryDto } from './dto/article-list-query.dto';
import { PaginatedArticlesResponse } from './dto/article-response.dto';
import { JwtAuthGuard } from '../../common/guards/jwt-auth.guard';
import { OptionalJwtAuthGuard } from '../../common/guards/optional-jwt-auth.guard';
import { RolesGuard } from '../../common/guards/roles.guard';
import { Roles } from '../../common/decorators/roles.decorator';
import { JwtPayload } from '../auth/interfaces/jwt-payload.interface';

@ApiTags('Articles')
@Controller('articles')
export class ArticlesController {
  constructor(private readonly articlesService: ArticlesService) {}

  // ─── PUBLIC: list articles (optional JWT for plan-level context) ───────────

  @Get()
  @UseGuards(OptionalJwtAuthGuard)
  @ApiOperation({
    summary: 'Danh sách bài viết',
    description: 'Danh sách bài viết đã xuất bản, phân trang cursor, lọc theo danh mục',
  })
  @ApiQuery({ name: 'category', required: false, description: 'Slug danh mục' })
  @ApiQuery({ name: 'cursor', required: false, description: 'ID bài viết cuối' })
  @ApiQuery({ name: 'limit', required: false, description: 'Số bài/trang (1-50)' })
  @ApiResponse({ status: 200, description: 'Danh sách bài viết phân trang' })
  async findAll(
    @Query() query: ArticleListQueryDto,
  ): Promise<PaginatedArticlesResponse> {
    return this.articlesService.findAll(query);
  }

  // ─── USER/EDITOR: list own articles ───────────────────────────────────────

  @Get('me')
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.admin, Role.editor, Role.user)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Danh sách bài viết của tôi',
    description: 'Lấy danh sách các bài viết do chính user hiện tại tạo.',
  })
  @ApiResponse({ status: 200, description: 'Danh sách bài viết phân trang' })
  @ApiResponse({ status: 401, description: 'Chưa đăng nhập' })
  async findOwn(
    @Query() query: ArticleListQueryDto,
    @Req() request: any,
  ): Promise<PaginatedArticlesResponse> {
    if (!request.user) {
      throw new Error('Authentication required');
    }
    return this.articlesService.findAllOwn(request.user.sub, query);
  }

  // ─── ADMIN: list all articles including drafts ────────────────────────────

  @Get('admin/list')
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.admin, Role.editor)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Danh sách bài viết (Admin)',
    description: 'Danh sách tất cả bài viết bao gồm bản nháp, phân trang cursor',
  })
  @ApiResponse({ status: 200, description: 'Danh sách bài viết phân trang' })
  async findAllAdmin(
    @Query() query: ArticleListQueryDto,
  ): Promise<PaginatedArticlesResponse> {
    return this.articlesService.findAllAdmin(query);
  }

  // ─── ADMIN: get article by slug including drafts ───────────────────────────

  @Get('admin/:slug')
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.admin, Role.editor)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Chi tiết bài viết (Admin)',
    description: 'Trả về bài viết đầy đủ bao gồm bản nháp cho admin/editor',
  })
  @ApiResponse({ status: 200, description: 'Chi tiết bài viết' })
  @ApiResponse({ status: 401, description: 'Chưa đăng nhập' })
  @ApiResponse({ status: 403, description: 'Không có quyền' })
  @ApiResponse({ status: 404, description: 'Không tìm thấy bài viết' })
  async findBySlugAdmin(@Param('slug') slug: string) {
    return this.articlesService.findBySlugAdmin(slug);
  }

  // ─── PUBLIC: get article by slug (optional JWT for plan gating) ────────────

  @Get(':slug')
  @UseGuards(OptionalJwtAuthGuard)
  @ApiOperation({
    summary: 'Chi tiết bài viết',
    description: 'Trả về bài viết đầy đủ. Nội dung premium yêu cầu gói phù hợp.',
  })
  @ApiResponse({ status: 200, description: 'Chi tiết bài viết' })
  @ApiResponse({ status: 403, description: 'Cần nâng cấp gói' })
  @ApiResponse({ status: 404, description: 'Không tìm thấy bài viết' })
  async findBySlug(
    @Param('slug') slug: string,
    @Req() request: any,
  ) {
    // request.user is typed via Fastify augmentation (see common/types/fastify.d.ts)
    const planLevel = request.user?.planLevel ?? 0;
    const userRole = request.user?.role;
    const bypassPlanCheck = userRole === 'admin' || userRole === 'editor';
    return this.articlesService.findBySlug(slug, planLevel, bypassPlanCheck);
  }

  // ─── ADMIN/EDITOR: create article ─────────────────────────────────────────

  @Post()
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.admin, Role.editor)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Tạo bài viết',
    description: 'Admin hoặc editor tạo bài viết mới',
  })
  @ApiResponse({ status: 201, description: 'Bài viết đã được tạo' })
  @ApiResponse({ status: 401, description: 'Chưa đăng nhập' })
  @ApiResponse({ status: 403, description: 'Không có quyền' })
  @ApiResponse({ status: 409, description: 'Slug đã tồn tại' })
  async create(@Body() dto: CreateArticleDto, @Req() request: any) {
    // JwtAuthGuard guarantees request.user is set
    if (!request.user) {
      throw new Error('Authentication required');
    }
    return this.articlesService.create(dto, request.user.sub);
  }

  // ─── ADMIN/EDITOR: update article ─────────────────────────────────────────

  @Patch(':slug')
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.admin, Role.editor)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Cập nhật bài viết',
    description: 'Cập nhật nội dung, metadata, hoặc phân quyền gói',
  })
  @ApiResponse({ status: 200, description: 'Bài viết đã được cập nhật' })
  @ApiResponse({ status: 401, description: 'Chưa đăng nhập' })
  @ApiResponse({ status: 403, description: 'Không có quyền' })
  @ApiResponse({ status: 404, description: 'Không tìm thấy bài viết' })
  @ApiResponse({ status: 409, description: 'Slug đã tồn tại' })
  async update(
    @Param('slug') slug: string,
    @Body() dto: UpdateArticleDto,
  ) {
    return this.articlesService.update(slug, dto);
  }

  // ─── ADMIN/AUTHOR: delete article ─────────────────────────────────────────

  @Delete(':slug')
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.admin, Role.editor, Role.user)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Xóa bài viết',
    description: 'Xóa vĩnh viễn bài viết. Admin hoặc tác giả của bài viết.',
  })
  @ApiResponse({ status: 200, description: 'Bài viết đã được xóa' })
  @ApiResponse({ status: 401, description: 'Chưa đăng nhập' })
  @ApiResponse({ status: 403, description: 'Không có quyền' })
  @ApiResponse({ status: 404, description: 'Không tìm thấy bài viết' })
  async remove(
    @Param('slug') slug: string,
    @Req() request: any,
  ): Promise<{ message: string }> {
    if (!request.user) {
      throw new Error('Authentication required');
    }
    await this.articlesService.remove(slug, request.user.sub, request.user.role);
    return { message: 'Bài viết đã được xóa' };
  }
}