# hotstock-be-v1/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/articles": {
      "get": {
        "description": "Danh sách bài viết đã xuất bản, phân trang cursor, lọc theo danh mục",
        "operationId": "ArticlesController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Slug danh mục",
            "schema": {
              "example": "phan-tich-ky-thuat",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "ID bài viết cuối",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số bài/trang (1-50)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách bài viết phân trang"
          }
        },
        "summary": "Danh sách bài viết",
        "tags": [
          "Articles"
        ]
      },
      "post": {
        "description": "Admin hoặc editor tạo bài viết mới",
        "operationId": "ArticlesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bài viết đã được tạo"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "409": {
            "description": "Slug đã tồn tại"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo bài viết",
        "tags": [
          "Articles"
        ]
      }
    },
    "/articles/me": {
      "get": {
        "description": "Lấy danh sách các bài viết do chính user hiện tại tạo.",
        "operationId": "ArticlesController_findOwn",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Lọc theo slug danh mục",
            "schema": {
              "example": "phan-tich-ky-thuat",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "ID bài viết cuối cùng (cursor pagination)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số bài viết mỗi trang (1-50, mặc định 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách bài viết phân trang"
          },
          "401": {
            "description": "Chưa đăng nhập"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Danh sách bài viết của tôi",
        "tags": [
          "Articles"
        ]
      }
    },
    "/articles/admin/list": {
      "get": {
        "description": "Danh sách tất cả bài viết bao gồm bản nháp, phân trang cursor",
        "operationId": "ArticlesController_findAllAdmin",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Lọc theo slug danh mục",
            "schema": {
              "example": "phan-tich-ky-thuat",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "ID bài viết cuối cùng (cursor pagination)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số bài viết mỗi trang (1-50, mặc định 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách bài viết phân trang"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Danh sách bài viết (Admin)",
        "tags": [
          "Articles"
        ]
      }
    },
    "/articles/admin/{slug}": {
      "get": {
        "description": "Trả về bài viết đầy đủ bao gồm bản nháp cho admin/editor",
        "operationId": "ArticlesController_findBySlugAdmin",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết bài viết"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Không tìm thấy bài viết"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Chi tiết bài viết (Admin)",
        "tags": [
          "Articles"
        ]
      }
    },
    "/articles/{slug}": {
      "get": {
        "description": "Trả về bài viết đầy đủ. Nội dung premium yêu cầu gói phù hợp.",
        "operationId": "ArticlesController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết bài viết"
          },
          "403": {
            "description": "Cần nâng cấp gói"
          },
          "404": {
            "description": "Không tìm thấy bài viết"
          }
        },
        "summary": "Chi tiết bài viết",
        "tags": [
          "Articles"
        ]
      },
      "patch": {
        "description": "Cập nhật nội dung, metadata, hoặc phân quyền gói",
        "operationId": "ArticlesController_update",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bài viết đã được cập nhật"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Không tìm thấy bài viết"
          },
          "409": {
            "description": "Slug đã tồn tại"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật bài viết",
        "tags": [
          "Articles"
        ]
      },
      "delete": {
        "description": "Xóa vĩnh viễn bài viết. Admin hoặc tác giả của bài viết.",
        "operationId": "ArticlesController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bài viết đã được xóa"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Không tìm thấy bài viết"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa bài viết",
        "tags": [
          "Articles"
        ]
      }
    },
...
  },
  "info": {
    "title": "HotStock API Documentation",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ArticleContentBlockDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "default-block"
          },
          "type": {
            "type": "string",
            "example": "text"
          },
          "content": {
            "type": "string",
            "example": "<p>Nội dung...</p>"
          }
        },
        "required": [
          "content"
        ]
      },
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Phân tích VN-Index tuần 24",
            "description": "Tiêu đề bài viết"
          },
          "description": {
            "type": "string",
            "example": "Tóm tắt nội dung bài viết",
            "description": "Mô tả ngắn"
          },
          "slug": {
            "type": "string",
            "example": "phan-tich-vn-index-tuan-24",
            "description": "Slug URL"
          },
          "publishedAt": {
            "type": "string",
            "example": "2024-06-15T00:00:00.000Z",
            "description": "Thời điểm xuất bản"
          },
          "contentBlocks": {
            "example": [
              {
                "id": "default-block",
                "type": "text",
                "content": "<p>Nội dung...</p>"
              }
            ],
            "description": "Nội dung bài viết dạng block",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleContentBlockDto"
            }
          },
          "coverUrl": {
            "type": "string",
            "example": "/uploads/cover.jpg"
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "description": "ID danh mục"
          },
          "tagIds": {
            "example": [
              1,
              2
            ],
            "description": "Mảng ID tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "readPermission": {
            "type": "string",
            "example": "public",
            "description": "Quyền truy cập (public, member, editor, admin)"
          },
          "planIds": {
            "example": [
              1,
              2
            ],
            "description": "Danh sách plan ID có quyền truy cập",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "slug",
          "contentBlocks",
          "categoryId"
        ]
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Phân tích VN-Index tuần 24",
            "description": "Tiêu đề bài viết"
          },
          "description": {
            "type": "string",
            "example": "Tóm tắt nội dung bài viết",
            "description": "Mô tả ngắn"
          },
          "slug": {
            "type": "string",
            "example": "phan-tich-vn-index-tuan-24",
            "description": "Slug URL"
          },
          "publishedAt": {
            "type": "string",
            "example": "2024-06-15T00:00:00.000Z",
            "description": "Thời điểm xuất bản"
          },
          "contentBlocks": {
            "example": [
              {
                "id": "default-block",
                "type": "text",
                "content": "<p>Nội dung...</p>"
              }
            ],
            "description": "Nội dung bài viết dạng block",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleContentBlockDto"
            }
          },
          "coverUrl": {
            "type": "string",
            "example": "/uploads/cover.jpg"
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "description": "ID danh mục"
          },
          "tagIds": {
            "example": [
              1,
              2
            ],
            "description": "Mảng ID tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "readPermission": {
            "type": "string",
            "example": "public",
            "description": "Quyền truy cập (public, member, editor, admin)"
          },
          "planIds": {
            "example": [
              1,
              2
            ],
            "description": "Danh sách plan ID có quyền truy cập",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}